Removing gcc warnings about signed vs unsigned and others
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6463 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/regex/regexec.c
@@ -180,7 +180,8 @@ int eflags;
 		return(REG_BADPAT);
 	eflags = GOODFLAGS(eflags);
 
-	if (g->nstates <= CHAR_BIT*sizeof(states1) && !(eflags&REG_LARGE))
+	if (g->nstates <= (int)(CHAR_BIT*sizeof(states1))
+	    && !(eflags&REG_LARGE))
 		return(smatcher(g, (char *)string, nmatch, pmatch, eflags));
 	else
 		return(lmatcher(g, (char *)string, nmatch, pmatch, eflags));
